Split the planting&harvest predictions into files
