+++ app/models/concerns/predict_harvest.rb
@@ -1,14 +1,7 @@
-module Predictable
+module PredictHarvest
   extend ActiveSupport::Concern
 
-  included do
-    ## Triggers
-    before_save :calculate_lifespan
-
-    def calculate_lifespan
-      self.lifespan = (planted_at.present? && finished_at.present? ? finished_at - planted_at : nil)
-    end
-
+  included do # rubocop:disable Metrics/BlockLength
     # dates
     def first_harvest_date
       harvests_with_dates.minimum(:harvested_at)
@@ -16,10 +9,6 @@ module Predictable
 
     def last_harvest_date
       harvests_with_dates.maximum(:harvested_at)
-    end
-
-    def finish_predicted_at
-      planted_at + crop.median_lifespan.days if crop.median_lifespan.present? && planted_at.present?
     end
 
     def first_harvest_predicted_at
@@ -30,31 +19,6 @@ module Predictable
     def last_harvest_predicted_at
       return unless crop.median_days_to_last_harvest.present? && planted_at.present?
       planted_at + crop.median_days_to_last_harvest.days
-    end
-
-    # days
-    def age_in_days
-      (Time.zone.today - planted_at).to_i if planted_at.present?
-    end
-
-    def expected_lifespan
-      if planted_at.present? && finished_at.present?
-        return (finished_at - planted_at).to_i
-      end
-      crop.median_lifespan
-    end
-
-    def days_since_planted
-      (Time.zone.today - planted_at).to_i if planted_at.present?
-    end
-
-    # progress
-    def percentage_grown
-      return 100 if finished
-      return if planted_at.blank? || expected_lifespan.blank?
-      p = (days_since_planted / expected_lifespan.to_f) * 100
-      return p if p <= 100
-      100
     end
 
     # actions
@@ -68,25 +32,7 @@ module Predictable
       update(days_to_first_harvest: days_to_first_harvest, days_to_last_harvest: days_to_last_harvest)
     end
 
-    # states
-    def finish_is_predicatable?
-      crop.annual? && planted_at.present? && finish_predicted_at.present?
-    end
-
-    def zombie?
-      crop.annual? && !finished &&
-        planted_at.present? &&
-        finish_predicted_at.present? &&
-        (finish_predicted_at + 60.days) < Time.zone.today
-    end
-
-    def should_be_finished?
-      crop.annual? && !finished &&
-        planted_at.present? &&
-        finish_predicted_at.present? &&
-        finish_predicted_at < Time.zone.today
-    end
-
+    # status
     def harvest_time?
       return false if crop.perennial || finished
 
@@ -104,6 +50,12 @@ module Predictable
         harvests.empty? &&
         first_harvest_predicted_at.present? &&
         first_harvest_predicted_at > Time.zone.today
+    end
+
+    private
+
+    def harvests_with_dates
+      harvests.where.not(harvested_at: nil)
     end
   end
 end
